Rename "medium touch" to "managed infrastructure" and "high touch" to "unmanaged infrastructure"
