+++ dev-setup/gardenadm/machines/ippool.yaml
@@ -1,7 +1,7 @@
 apiVersion: crd.projectcalico.org/v1
 kind: IPPool
 metadata:
-  name: gardenadm-high-touch
+  name: gardenadm-unmanaged-infra
 spec:
   cidr: 10.0.0.0/16
   ipipMode: Always
