+++ dev-setup/gardenadm/machines/namespace.yaml
@@ -1,7 +1,7 @@
 apiVersion: v1
 kind: Namespace
 metadata:
-  name: gardenadm-high-touch
+  name: gardenadm-unmanaged-infra
   labels:
     # Use label so that machine pods are considered in the export_artifacts() helper function during test runs in Prow.
     # This makes sure that the usual systemd and pod logs get exported for each machine pod in this namespace.
