+++ dev-setup/gardenadm/resources/overlays/managed-infra/patch-shoot-managed-infra.yaml
