+++ pkg/apis/core/v1beta1/helper/shoot.go
@@ -644,7 +644,6 @@ func IsShootSelfHosted(workers []gardencorev1beta1.Worker) bool {
 }
 
 // HasManagedInfrastructure returns true if the shoot's infrastructure (network, machines, etc.) is managed by Gardener.
-// I.e., it returns false for high-touch self-hosted shoots, where the infrastructure is managed by the user.
 func HasManagedInfrastructure(shoot *gardencorev1beta1.Shoot) bool {
 	return shoot.Spec.CredentialsBindingName != nil || shoot.Spec.SecretBindingName != nil
 }
