+++ test/e2e/gardenadm/e2e_suite_test.go
@@ -14,8 +14,8 @@ import (
 
 	"github.com/gardener/gardener/pkg/logger"
 	. "github.com/gardener/gardener/test/e2e/gardenadm/common"
-	_ "github.com/gardener/gardener/test/e2e/gardenadm/hightouch"
-	_ "github.com/gardener/gardener/test/e2e/gardenadm/mediumtouch"
+	_ "github.com/gardener/gardener/test/e2e/gardenadm/managedinfra"
+	_ "github.com/gardener/gardener/test/e2e/gardenadm/unmanagedinfra"
 )
 
 func TestE2E(t *testing.T) {
